[visualization] Refactor ModelVisualization to fit new interface style
