+++ src/visualization/src/ModelVisualization.cpp
@@ -1,0 +1,215 @@
+/*
+ * Copyright (C) 2016 Fondazione Istituto Italiano di Tecnologia
+ * Authors: Silvio Traversaro
+ * CopyPolicy: Released under the terms of the LGPLv2.1 or later, see LGPL.TXT
+ *
+ */
+
+#include "ModelVisualization.h"
+#include "IrrlichtUtils.h"
+
+#include <iDynTree/Model/ForwardKinematics.h>
+#include <iDynTree/Model/Model.h>
+#include <iDynTree/Model/Traversal.h>
+
+namespace iDynTree
+{
+
+struct ModelVisualization::ModelVisualizationPimpl
+{
+    std::string m_instanceName;
+
+    /**
+     * Set to true if is a valid instance, false otherwise.
+     */
+    bool m_isValid;
+
+    /**
+     * Used model.
+     */
+    Model m_model;
+
+    /**
+     * Used traversal.
+     */
+    Traversal m_traversal;
+
+    /**
+     * Buffer for forward kinematics.
+     */
+    LinkPositions m_fwdKinBuffer;
+
+    irr::scene::ISceneNode * modelNode;
+    std::vector<irr::scene::ISceneNode *> linkNodes;
+    std::vector<irr::scene::ISceneNode *> linkFramesNodes;
+    std::vector< std::vector<irr::scene::ISceneNode *> > geomNodes;
+    irr::scene::ISceneManager * m_irrSmgr;
+
+    void addModelGeometriesToSceneManager(const iDynTree::Model & model, const iDynTree::ModelSolidShapes & modelGeom);
+    void updateLinkPositions(const iDynTree::LinkPositions & world_H_link);
+
+
+    ModelVisualizationPimpl()
+    {
+        m_isValid = false;
+        m_instanceName = "";
+    }
+};
+
+
+/**
+ * Add  iDynTree::ModelGeometries to an irr::scene::ISceneManager*
+ * We create a SceneGraph for the URDF model of this type: the root object
+ * is the model, its child are the links, and the child of the links are the actual geometric objects
+ * Note that in this case all the links are child of the model, and the scene graph does not mirror
+ *  the kinematic graph of iDynTree Model
+ */
+void ModelVisualization::ModelVisualizationPimpl::addModelGeometriesToSceneManager(const iDynTree::Model & model,
+                                                                                   const iDynTree::ModelSolidShapes & modelGeom)
+{
+    this->modelNode = this->m_irrSmgr->addEmptySceneNode();
+    this->linkNodes.resize(model.getNrOfLinks());
+    this->linkFramesNodes.resize(model.getNrOfLinks());
+    this->geomNodes.resize(model.getNrOfLinks());
+
+    for(size_t linkIdx=0; linkIdx < model.getNrOfLinks(); linkIdx++)
+    {
+        this->linkNodes[linkIdx] = this->m_irrSmgr->addEmptySceneNode(this->modelNode);
+
+        this->geomNodes[linkIdx].resize(modelGeom.linkSolidShapes[linkIdx].size());
+
+        for(size_t geom=0; geom < modelGeom.linkSolidShapes[linkIdx].size(); geom++)
+        {
+            this->geomNodes[linkIdx][geom] = addGeometryToSceneManager(modelGeom.linkSolidShapes[linkIdx][geom],this->linkNodes[linkIdx],this->m_irrSmgr);
+        }
+    }
+}
+
+void ModelVisualization::ModelVisualizationPimpl::updateLinkPositions(const iDynTree::LinkPositions & world_H_link)
+{
+    for(size_t linkIdx=0; linkIdx < world_H_link.getNrOfLinks(); linkIdx++)
+    {
+        setWorldHNode(this->linkNodes[linkIdx],world_H_link(linkIdx));
+    }
+}
+
+ModelVisualization::ModelVisualization():
+    pimpl(new ModelVisualizationPimpl())
+{
+}
+
+ModelVisualization::~ModelVisualization()
+{
+    this->close();
+    if( pimpl )
+    {
+        delete pimpl;
+        pimpl = 0;
+    }
+}
+
+
+ModelVisualization::ModelVisualization(const ModelVisualization& /*other*/)
+{
+    assert(false);
+}
+
+ModelVisualization& ModelVisualization::operator=(const ModelVisualization& /*other*/)
+{
+    assert(false);
+    return *this;
+}
+
+bool ModelVisualization::init(const Model& model,
+                              const std::string instanceName,
+                              irr::scene::ISceneManager * sceneManager)
+{
+    // Check if the visual of the models are consisten with the rest of the model
+    if( !model.visualSolidShapes().isConsistent(model) )
+    {
+        reportError("ModelVisualization","init","Impossible to use load model, as the visual solid shapes of the model are not consistent with the model itself.");
+        return false;
+    }
+    this->pimpl->m_instanceName = instanceName;
+
+    this->pimpl->m_irrSmgr = sceneManager;
+
+    // Copy the model and create a traversal from the default base
+    this->pimpl->m_model = model;
+    this->pimpl->m_model.computeFullTreeTraversal(this->pimpl->m_traversal);
+
+    // Resize fwd kinematics buffer
+    this->pimpl->m_fwdKinBuffer.resize(this->pimpl->m_model);
+
+    // Create model in the scene, using visual solidShapes
+    pimpl->addModelGeometriesToSceneManager(model,model.visualSolidShapes());
+
+    // Set the initial position of the model
+    Transform world_H_base = Transform::Identity();
+    JointPosDoubleArray jointPos(model);
+    jointPos.zero();
+
+    this->setPositions(world_H_base,jointPos);
+
+    pimpl->m_isValid = true;
+
+    return true;
+}
+
+Model& ModelVisualization::model()
+{
+    return this->pimpl->m_model;
+}
+
+bool ModelVisualization::setPositions(const Transform& world_H_base, const VectorDynSize& jointPos)
+{
+    if( (jointPos.size() != model().getNrOfPosCoords()) )
+    {
+        std::stringstream ss;
+        ss << "Input size mismatch: model internal position coords " << model().getNrOfPosCoords() << " provided vector " << jointPos.size();
+        reportError("ModelVisualization","setPositions",ss.str().c_str());
+        return false;
+    }
+
+    // Compute fwd kinematics
+    bool ok = ForwardPositionKinematics(model(), this->pimpl->m_traversal,
+                                        world_H_base, jointPos,
+                                        this->pimpl->m_fwdKinBuffer);
+
+    if( ok )
+    {
+        this->pimpl->updateLinkPositions(this->pimpl->m_fwdKinBuffer);
+    }
+    else
+    {
+        reportError("ModelVisualization","setPositions","Forward kinematics error.");
+    }
+
+    return ok;
+}
+
+bool ModelVisualization::setLinkPositions(const LinkPositions& linkPos)
+{
+    if( !linkPos.isConsistent(model()) )
+    {
+        reportError("ModelVisualization","setLinkPositions","Input size mismatch.");
+        return false;
+    }
+
+    this->pimpl->updateLinkPositions(linkPos);
+    return true;
+}
+
+std::string ModelVisualization::getInstanceName()
+{
+    return this->pimpl->m_instanceName;
+}
+
+
+void ModelVisualization::close()
+{
+
+}
+
+
+}
