+++ src/visualization/src/Visualizer.cpp
@@ -6,9 +6,6 @@
  */
 
 #include <iDynTree/Core/Utils.h>
-#include <iDynTree/Model/ForwardKinematics.h>
-#include <iDynTree/Model/Model.h>
-#include <iDynTree/Model/Traversal.h>
 #include <iDynTree/Model/JointState.h>
 #include <iDynTree/Visualizer.h>
 
@@ -17,6 +14,7 @@
 #include "IrrlichtUtils.h"
 #include "Camera.h"
 #include "Environment.h"
+#include "ModelVisualization.h"
 #endif
 
 #include "DummyImplementations.h"
@@ -31,6 +29,10 @@ ICamera::~ICamera()
 }
 
 IEnvironment::~IEnvironment()
+{
+}
+
+IModelVisualization::~IModelVisualization()
 {
 }
 
@@ -52,48 +54,6 @@ ColorViz::ColorViz(const Vector4& rgba): r(rgba(0)), g(rgba(1)), b(rgba(2)), a(r
 
 }
 
-struct ModelVisualization::ModelVisualizationPimpl
-{
-    std::string m_instanceName;
-
-    /**
-     * Set to true if is a valid instance, false otherwise.
-     */
-    bool m_isValid;
-
-    /**
-     * Used model.
-     */
-    Model m_model;
-
-    /**
-     * Used traversal.
-     */
-    Traversal m_traversal;
-
-    /**
-     * Buffer for forward kinematics.
-     */
-    LinkPositions m_fwdKinBuffer;
-
-#ifdef IDYNTREE_USES_IRRLICHT
-    irr::scene::ISceneNode * modelNode;
-    std::vector<irr::scene::ISceneNode *> linkNodes;
-    std::vector<irr::scene::ISceneNode *> linkFramesNodes;
-    std::vector< std::vector<irr::scene::ISceneNode *> > geomNodes;
-    irr::scene::ISceneManager * m_irrSmgr;
-
-    void addModelGeometriesToSceneManager(const iDynTree::Model & model, const iDynTree::ModelSolidShapes & modelGeom);
-    void updateLinkPositions(const iDynTree::LinkPositions & world_H_link);
-#endif
-
-    ModelVisualizationPimpl()
-    {
-        m_isValid = false;
-        m_instanceName = "";
-    }
-};
-
 struct Visualizer::VisualizerPimpl
 {
     /**
@@ -109,7 +69,7 @@ struct Visualizer::VisualizerPimpl
     /**
      * Invalid model visualization, useful to return in case of error.
      */
-    ModelVisualization m_invalidModelViz;
+    DummyModelVisualization m_invalidModelViz;
 
     /**
      * Last FPS measured.
@@ -160,187 +120,6 @@ struct Visualizer::VisualizerPimpl
 #endif
     }
 };
-
-
-#ifdef IDYNTREE_USES_IRRLICHT
-
-/**
- * Add  iDynTree::ModelGeometries to an irr::scene::ISceneManager*
- * We create a SceneGraph for the URDF model of this type: the root object
- * is the model, its child are the links, and the child of the links are the actual geometric objects
- * Note that in this case all the links are child of the model, and the scene graph does not mirror
- *  the kinematic graph of iDynTree Model
- */
-void ModelVisualization::ModelVisualizationPimpl::addModelGeometriesToSceneManager(const iDynTree::Model & model,
-                                                                                   const iDynTree::ModelSolidShapes & modelGeom)
-{
-    this->modelNode = this->m_irrSmgr->addEmptySceneNode();
-    this->linkNodes.resize(model.getNrOfLinks());
-    this->linkFramesNodes.resize(model.getNrOfLinks());
-    this->geomNodes.resize(model.getNrOfLinks());
-
-    for(size_t linkIdx=0; linkIdx < model.getNrOfLinks(); linkIdx++)
-    {
-        this->linkNodes[linkIdx] = this->m_irrSmgr->addEmptySceneNode(this->modelNode);
-
-        this->geomNodes[linkIdx].resize(modelGeom.linkSolidShapes[linkIdx].size());
-
-        for(size_t geom=0; geom < modelGeom.linkSolidShapes[linkIdx].size(); geom++)
-        {
-            this->geomNodes[linkIdx][geom] = addGeometryToSceneManager(modelGeom.linkSolidShapes[linkIdx][geom],this->linkNodes[linkIdx],this->m_irrSmgr);
-        }
-    }
-}
-
-void ModelVisualization::ModelVisualizationPimpl::updateLinkPositions(const iDynTree::LinkPositions & world_H_link)
-{
-    for(size_t linkIdx=0; linkIdx < world_H_link.getNrOfLinks(); linkIdx++)
-    {
-        setWorldHNode(this->linkNodes[linkIdx],world_H_link(linkIdx));
-    }
-}
-
-#endif
-
-ModelVisualization::ModelVisualization():
-    pimpl(new ModelVisualizationPimpl())
-{
-}
-
-ModelVisualization::~ModelVisualization()
-{
-    this->close();
-    if( pimpl )
-    {
-        delete pimpl;
-        pimpl = 0;
-    }
-}
-
-
-ModelVisualization::ModelVisualization(const ModelVisualization& /*other*/)
-{
-    assert(false);
-}
-
-ModelVisualization& ModelVisualization::operator=(const ModelVisualization& /*other*/)
-{
-    assert(false);
-    return *this;
-}
-
-bool ModelVisualization::init(const Model& model,
-                              const std::string instanceName,
-                              Visualizer& visualizer)
-{
-#ifdef IDYNTREE_USES_IRRLICHT
-    // Check if the visual of the models are consisten with the rest of the model
-    if( !model.visualSolidShapes().isConsistent(model) )
-    {
-        reportError("ModelVisualization","init","Impossible to use load model, as the visual solid shapes of the model are not consistent with the model itself.");
-        return false;
-    }
-    this->pimpl->m_instanceName = instanceName;
-
-    this->pimpl->m_irrSmgr = visualizer.pimpl->m_irrSmgr;
-
-    // Copy the model and create a traversal from the default base
-    this->pimpl->m_model = model;
-    this->pimpl->m_model.computeFullTreeTraversal(this->pimpl->m_traversal);
-
-    // Resize fwd kinematics buffer
-    this->pimpl->m_fwdKinBuffer.resize(this->pimpl->m_model);
-
-    // Create model in the scene, using visual solidShapes
-    pimpl->addModelGeometriesToSceneManager(model,model.visualSolidShapes());
-
-    // Set the initial position of the model
-    Transform world_H_base = Transform::Identity();
-    JointPosDoubleArray jointPos(model);
-    jointPos.zero();
-
-    this->setPositions(world_H_base,jointPos);
-
-    pimpl->m_isValid = true;
-
-    return true;
-#else
-    IDYNTREE_UNUSED(model);
-    IDYNTREE_UNUSED(instanceName);
-    IDYNTREE_UNUSED(visualizer);
-    reportError("ModelVisualization","init","Impossible to use iDynTree::Visualizer, as iDynTree has been compiled without Irrlicht.");
-    return false;
-#endif
-}
-
-Model& ModelVisualization::model()
-{
-    return this->pimpl->m_model;
-}
-
-bool ModelVisualization::setPositions(const Transform& world_H_base, const VectorDynSize& jointPos)
-{
-#ifdef IDYNTREE_USES_IRRLICHT
-    if( (jointPos.size() != model().getNrOfPosCoords()) )
-    {
-        std::stringstream ss;
-        ss << "Input size mismatch: model internal position coords " << model().getNrOfPosCoords() << " provided vector " << jointPos.size();
-        reportError("ModelVisualization","setPositions",ss.str().c_str());
-        return false;
-    }
-
-    // Compute fwd kinematics
-    bool ok = ForwardPositionKinematics(model(), this->pimpl->m_traversal,
-                                        world_H_base, jointPos,
-                                        this->pimpl->m_fwdKinBuffer);
-
-    if( ok )
-    {
-        this->pimpl->updateLinkPositions(this->pimpl->m_fwdKinBuffer);
-    }
-    else
-    {
-        reportError("ModelVisualization","setPositions","Forward kinematics error.");
-    }
-
-    return ok;
-
-#else
-    IDYNTREE_UNUSED(world_H_base);
-    IDYNTREE_UNUSED(jointPos);
-    reportError("ModelVisualization","setPositions","Impossible to use iDynTree::Visualizer, as iDynTree has been compiled without Irrlicht.");
-    return false;
-#endif
-}
-
-bool ModelVisualization::setLinkPositions(const LinkPositions& linkPos)
-{
-#ifdef IDYNTREE_USES_IRRLICHT
-    if( !linkPos.isConsistent(model()) )
-    {
-        reportError("ModelVisualization","setLinkPositions","Input size mismatch.");
-        return false;
-    }
-
-    this->pimpl->updateLinkPositions(linkPos);
-    return true;
-#else
-    IDYNTREE_UNUSED(linkPos);
-    reportError("ModelVisualization","setLinkPositions","Impossible to use iDynTree::Visualizer, as iDynTree has been compiled without Irrlicht.");
-    return false;
-#endif
-}
-
-std::string ModelVisualization::getInstanceName()
-{
-    return this->pimpl->m_instanceName;
-}
-
-
-void ModelVisualization::close()
-{
-
-}
 
 
 Visualizer::Visualizer(const Visualizer& /*other*/)
@@ -491,7 +270,7 @@ bool Visualizer::addModel(const Model& model, const std::string& instanceName)
 
     ModelVisualization * newModelViz = new ModelVisualization();
 
-    if( !newModelViz->init(model,instanceName,*this) )
+    if( !newModelViz->init(model,instanceName,pimpl->m_irrSmgr) )
     {
         delete newModelViz;
         return false;
@@ -585,7 +364,7 @@ bool Visualizer::drawToFile(const std::string filename)
 }
 
 
-ModelVisualization& Visualizer::modelViz(const std::string& instanceName)
+IModelVisualization& Visualizer::modelViz(const std::string& instanceName)
 {
     int idx = getModelInstanceIndex(instanceName);
     if( idx < 0 )
@@ -596,7 +375,7 @@ ModelVisualization& Visualizer::modelViz(const std::string& instanceName)
     return *(this->pimpl->m_modelViz[idx]);
 }
 
-ModelVisualization& Visualizer::modelViz(size_t modelIdx)
+IModelVisualization& Visualizer::modelViz(size_t modelIdx)
 {
     return *(this->pimpl->m_modelViz[modelIdx]);
 }
